fix(dashboard): put the workspace picker on the tokens, and stop doubling an error
Review of #626, two fixes to the organization roster.

The workspace assignment list used a bare `<input type="checkbox">`, which
carries no styling at all, so it rendered as an unstyled browser control next to
token-styled everything else, with a hit target around 13px. The frontend
standards rule out hand-rolled primitives, and the table already had the answer:
`DataTable`'s selection box visual moves to `components/ui.tsx` as
`CheckboxVisual`, and a `Checkbox` beside it wraps it for standalone use. One
visual serves both, so a checkbox in a form and a table's selection box cannot
drift apart. The three bare checkboxes on `KeysPage`, `SettingsPage` and
`UsersPage` predate this PR and are a separate cleanup, but the component they
need is now there.

The page also listed its remove mutation's error in the banner and again as the
confirm dialog's own error, so a refused removal showed the same message twice,
once behind the open dialog. The dialog owns that mutation's error; the banner
covers the reads and the inline update.
